resolves #50 cleaning up flask server, still some delay issues running from command line, will open a new ticket for that
